# internal/server/openapi.yaml
openapi: 3.1.0
info:
  title: Kino HTTP API
  summary: Stable, versioned contract for the Kino ROM library
  version: 0.2.0-preview.1
  description: |
    ROM catalog, metadata imports, managed media, and library maintenance.
    Protected operations require the configured owner bearer token.
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
...
servers:
  - url: /api/v1
    description: Same-origin stable API entry point
security:
  - bearerAuth: []
...
paths:
...
  /platforms:
    get:
      operationId: getPlatforms
      tags: [Platforms]
      summary: List platform presets
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Offset'}
        - {name: q, in: query, schema: {type: string}, description: "Search ID, localized name, vendor, aliases, and frontend system names."}
        - {name: category, in: query, schema: {type: string, enum: [console, handheld, arcade, computer]}}
      responses:
        '200': {description: Paginated platforms, content: {application/json: {schema: {$ref: '#/components/schemas/PlatformCollection'}}}}
        '400': {$ref: '#/components/responses/Error'}
  /platforms/{id}:
    get:
      operationId: getPlatformsById
      tags: [Platforms]
      summary: Resolve a platform ID or alias
      parameters: [{$ref: '#/components/parameters/ID'}]
      responses:
        '200': {description: Platform preset, content: {application/json: {schema: {$ref: '#/components/schemas/Platform'}}}}
        '404': {$ref: '#/components/responses/Error'}
  /custom-platforms:
    get:
      operationId: getCustomPlatforms
      tags: [Platforms]
      summary: List user-defined platform records, including disabled entries
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Offset'}]
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Custom platforms, content: {application/json: {schema: {$ref: '#/components/schemas/CustomPlatformCollection'}}}}
    post:
      operationId: postCustomPlatforms
      tags: [Platforms]
      summary: Create a custom platform with audited aliases, ROM extensions, and frontend mappings
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/NewCustomPlatform'}}}}
      responses:
        '201': {description: Created custom platform, content: {application/json: {schema: {$ref: '#/components/schemas/CustomPlatform'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /custom-platforms/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get:
      operationId: getCustomPlatformsById
      tags: [Platforms]
      summary: Get a custom platform
      responses:
        '200': {description: Custom platform, content: {application/json: {schema: {$ref: '#/components/schemas/CustomPlatform'}}}}
        '404': {$ref: '#/components/responses/Error'}
    put:
      operationId: putCustomPlatformsById
      tags: [Platforms]
      summary: Replace a custom platform without changing its ID
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/NewCustomPlatform'}}}}
      responses:
        '200': {description: Updated custom platform, content: {application/json: {schema: {$ref: '#/components/schemas/CustomPlatform'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
    delete:
      operationId: deleteCustomPlatformsById
      tags: [Platforms]
      summary: Delete an unused custom platform; referenced platforms must be disabled instead
      responses:
        '204': {description: Deleted}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /series:
    get:
      operationId: getSeries
      tags: [Series]
      summary: List cross-platform series with member Games
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Offset'}
        - {$ref: '#/components/parameters/Locale'}
        - {$ref: '#/components/parameters/Projection'}
        - {name: q, in: query, schema: {type: string}, description: "Search series titles, descriptions, member titles, platforms, and relation types."}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Paginated full or summary series, content: {application/json: {schema: {oneOf: [{$ref: '#/components/schemas/SeriesCollection'}, {$ref: '#/components/schemas/SeriesSummaryCollection'}]}}}}
    post:
      operationId: postSeries
      tags: [Series]
      summary: Create a cross-platform series and its optional complete member set atomically
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SeriesMutation'}
      responses:
        '201': {description: Created series, content: {application/json: {schema: {$ref: '#/components/schemas/Series'}}}}
        '400': {$ref: '#/components/responses/Error'}
  /series/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get:
      operationId: getSeriesById
      tags: [Series]
      summary: Get a series and its fully expanded member Games
      parameters: [{$ref: '#/components/parameters/Locale'}]
      responses:
        '200': {description: Series, content: {application/json: {schema: {$ref: '#/components/schemas/Series'}}}}
        '404': {$ref: '#/components/responses/Error'}
    put:
      operationId: putSeriesById
      tags: [Series]
      summary: Replace editable series metadata and, when supplied, its complete member set atomically
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SeriesMutation'}
      responses:
        '200': {description: Updated series, content: {application/json: {schema: {$ref: '#/components/schemas/Series'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
    delete:
      operationId: deleteSeriesById
      tags: [Series]
      summary: Delete only the presentation grouping; member Games and files remain intact
      responses:
        '204': {description: Deleted}
        '404': {$ref: '#/components/responses/Error'}
  /series/{id}/members/{game_id}:
    parameters:
      - {$ref: '#/components/parameters/ID'}
      - {name: game_id, in: path, required: true, schema: {type: string, format: uuid}}
    put:
      operationId: putSeriesByIdMembersByGameId
      tags: [Series]
      summary: Add or update a Game membership; Games from any platform are accepted
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                relation_type: {type: string, enum: [mainline, port, remake, spinoff, collection, other], default: mainline}
                sort_order: {type: integer, minimum: 0, default: 0}
      responses:
        '200': {description: Updated series, content: {application/json: {schema: {$ref: '#/components/schemas/Series'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
    delete:
      operationId: deleteSeriesByIdMembersByGameId
      tags: [Series]
      summary: Remove only the membership; Game, Editions, ROMs, and saves remain intact
      responses:
        '204': {description: Membership removed}
        '404': {$ref: '#/components/responses/Error'}
  /games:
    get:
      operationId: getGames
      tags: [Library]
      summary: List games with editions and artifacts
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Offset'}
        - {$ref: '#/components/parameters/Locale'}
        - {$ref: '#/components/parameters/Projection'}
        - {name: q, in: query, schema: {type: string}}
        - {name: platform, in: query, schema: {type: string}}
      responses:
        '200': {description: Paginated full or summary games, content: {application/json: {schema: {oneOf: [{$ref: '#/components/schemas/GameCollection'}, {$ref: '#/components/schemas/GameSummaryCollection'}]}}}}
        '400': {$ref: '#/components/responses/Error'}
    post:
      operationId: postGames
      tags: [Library]
      summary: Create a game
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/GameMutation'}}}}
      responses:
        '201': {description: Created game, content: {application/json: {schema: {$ref: '#/components/schemas/Game'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/ImportConflict'}
  /games/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get:
      operationId: getGamesById
      tags: [Library]
      summary: Get a game
      parameters: [{$ref: '#/components/parameters/Locale'}]
      responses:
        '200': {description: Game, content: {application/json: {schema: {$ref: '#/components/schemas/Game'}}}}
        '404': {$ref: '#/components/responses/Error'}
    put:
      operationId: putGamesById
      tags: [Library]
      summary: Replace editable game metadata
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/GameMutation'}}}}
      responses:
        '200': {description: Updated game, content: {application/json: {schema: {$ref: '#/components/schemas/Game'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
    delete:
      operationId: deleteGamesById
      tags: [Library]
      summary: Delete catalog metadata without deleting ROM files
      responses:
        '204': {description: Deleted}
        '404': {$ref: '#/components/responses/Error'}
  /games/{id}/merge:
    post:
      operationId: postGamesByIdMerge
      tags: [Library]
      summary: Merge another same-platform game into this game
      description: Current clients first obtain a signed preview and submit its token and fingerprint. Omitting both remains supported for legacy clients; supplying only one is rejected without writes. A changed reviewed graph returns game_merge_stale with zero writes.
      parameters: [{$ref: '#/components/parameters/ID'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/GameMergeCommitRequest'}}}
      responses:
        '200': {description: Merged game, content: {application/json: {schema: {$ref: '#/components/schemas/Game'}}}}
        '409': {$ref: '#/components/responses/Error'}
  /games/{id}/merge/preview:
    post:
      operationId: postGamesByIdMergePreview
      tags: [Library]
      summary: Preview and sign an atomic same-platform game merge
      description: Returns counts and invariants without exposing ROM paths or hashes. The fingerprint covers both games, localized names, editions, artifacts, media, and series membership.
      parameters:
        - {$ref: '#/components/parameters/ID'}
        - {$ref: '#/components/parameters/Locale'}
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/GameMergeRequest'}}}
      responses:
        '200': {description: Signed merge preview, content: {application/json: {schema: {$ref: '#/components/schemas/GameMergePreview'}}}}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /games/{id}/primary:
    put:
      operationId: putGamesByIdPrimary
      tags: [Library]
      summary: Set the primary edition
      parameters: [{$ref: '#/components/parameters/ID'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PrimaryEditionRequest'}}}}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Updated game, content: {application/json: {schema: {$ref: '#/components/schemas/Game'}}}}
  /editions:
    post:
      operationId: postEditions
      tags: [Library]
      summary: Create an independent game edition
      description: When artifact_path is supplied, the file is resolved below the library root and hashed before the edition and first artifact are committed in one transaction. A missing or unreadable path creates neither record. Leave artifact_path empty for an explicit metadata-only edition.
      requestBody: {$ref: '#/components/requestBodies/EditionCreateRequest'}
      responses:
        '201': {description: Game containing the new edition, content: {application/json: {schema: {oneOf: [{$ref: '#/components/schemas/Game'}, {$ref: '#/components/schemas/Edition'}]}}}}
        '400': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
  /editions/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get:
      operationId: getEditionsById
      tags: [Library]
      summary: Get an edition
      parameters: [{$ref: '#/components/parameters/Locale'}]
      responses:
        '200': {description: Edition, content: {application/json: {schema: {$ref: '#/components/schemas/Edition'}}}}
        '404': {$ref: '#/components/responses/Error'}
    put:
      operationId: putEditionsById
      tags: [Library]
      summary: Replace editable edition metadata
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EditionMutation'}}}}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Updated edition, content: {application/json: {schema: {$ref: '#/components/schemas/Edition'}}}}
    delete:
      operationId: deleteEditionsById
      tags: [Library]
      summary: Delete edition metadata without deleting ROM files
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '204': {description: Deleted}
  /editions/{id}/move:
    post:
      operationId: postEditionsByIdMove
      tags: [Library]
      summary: Move an edition to another same-platform game
      parameters: [{$ref: '#/components/parameters/ID'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EditionMoveRequest'}}}}
      responses:
        '200': {description: Moved edition, content: {application/json: {schema: {$ref: '#/components/schemas/Edition'}}}}
        '409': {$ref: '#/components/responses/Error'}
  /artifacts:
    post:
      operationId: postArtifacts
      tags: [Library]
      summary: Associate a ROM, disc, directory, patch, update, or DLC path
      description: The path must already exist below the configured library root and be hashable. This endpoint only creates a catalog relationship; it never moves, renames, overwrites, or deletes the source.
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ArtifactCreateRequest'}}}}
      responses:
        '201': {description: Created artifact, content: {application/json: {schema: {$ref: '#/components/schemas/Artifact'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
  /artifacts/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get:
      operationId: getArtifactsById
      tags: [Library]
      summary: Get an artifact
      responses:
        '200': {description: Artifact, content: {application/json: {schema: {$ref: '#/components/schemas/Artifact'}}}}
        '404': {$ref: '#/components/responses/Error'}
    put:
      operationId: putArtifactsById
      tags: [Library]
      summary: Update and recheck an artifact
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ArtifactUpdateRequest'}}}}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Updated artifact, content: {application/json: {schema: {$ref: '#/components/schemas/Artifact'}}}}
    delete:
      operationId: deleteArtifactsById
      tags: [Library]
      summary: Remove the association without deleting the file
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '204': {description: Deleted}
  /artifacts/recheck:
    post:
      operationId: postArtifactsRecheck
      tags: [Library]
      summary: Recheck all artifact paths and hashes
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Recheck counts, content: {application/json: {schema: {$ref: '#/components/schemas/ArtifactRecheckResult'}}}}
  /import-sources:
    get:
      operationId: getImportSources
      tags: [Transfer]
      summary: Discover frontend metadata inside the library root
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Offset'}
        - {name: format, in: query, required: true, schema: {type: string, enum: [pegasus, es-de]}}
      responses:
        '200': {description: Paginated import sources, content: {application/json: {schema: {$ref: '#/components/schemas/ImportSourceCollection'}}}}
        '400': {$ref: '#/components/responses/Error'}
  /sources:
    get:
      operationId: getSources
      tags: [Transfer]
      summary: List durable library source configurations
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Offset'}]
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Paginated library sources, content: {application/json: {schema: {$ref: '#/components/schemas/LibrarySourceCollection'}}}}
    post:
      operationId: postSources
      tags: [Transfer]
      summary: Save a read-only source configuration below the library root
      requestBody: {$ref: '#/components/requestBodies/LibrarySourceRequest'}
      responses:
        '201': {description: Created source, content: {application/json: {schema: {$ref: '#/components/schemas/LibrarySource'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /sources/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get:
      operationId: getSourcesById
      tags: [Transfer]
      summary: Get a durable library source
      responses:
        '200': {description: Library source, content: {application/json: {schema: {$ref: '#/components/schemas/LibrarySource'}}}}
        '404': {$ref: '#/components/responses/Error'}
    put:
      operationId: putSourcesById
      tags: [Transfer]
      summary: Replace source settings without changing any source file
      requestBody: {$ref: '#/components/requestBodies/LibrarySourceRequest'}
      responses:
        '200': {description: Updated source, content: {application/json: {schema: {$ref: '#/components/schemas/LibrarySource'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
    delete:
      operationId: deleteSourcesById
      tags: [Transfer]
      summary: Delete configuration metadata only when it has no scan history
      description: Source files are never deleted. Disable a source to retain its audit history.
      responses:
        '204': {description: Configuration deleted; source files unchanged}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
  /sources/{id}/scans:
    parameters: [{$ref: '#/components/parameters/ID'}]
    post:
      operationId: postSourcesByIdScans
      tags: [Transfer]
      summary: Run a read-only scan and persist its bounded preview audit
      responses:
        '201': {description: Signed source preview, content: {application/json: {schema: {$ref: '#/components/schemas/SourceScanPreview'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
  /source-scans:
    get:
      operationId: getSourceScans
      tags: [Transfer]
      summary: List source scan audit records
      parameters:
        - {$ref: '#/components/parameters/Limit'}
        - {$ref: '#/components/parameters/Offset'}
        - {name: source_id, in: query, schema: {type: string, format: uuid}}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Paginated source scans, content: {application/json: {schema: {$ref: '#/components/schemas/SourceScanCollection'}}}}
  /source-scans/{id}:
    parameters: [{$ref: '#/components/parameters/ID'}]
    get:
      operationId: getSourceScansById
      tags: [Transfer]
      summary: Get one source scan audit record
      responses:
        '200': {description: Source scan, content: {application/json: {schema: {$ref: '#/components/schemas/SourceScan'}}}}
        '404': {$ref: '#/components/responses/Error'}
  /source-scans/{id}/commit:
    parameters: [{$ref: '#/components/parameters/ID'}]
    post:
      operationId: postSourceScansByIdCommit
      tags: [Transfer]
      summary: Rescan and atomically commit a non-expired persistent preview
      requestBody: {$ref: '#/components/requestBodies/SourceScanCommitRequest'}
      responses:
        '200': {description: Atomic import result plus committed scan, content: {application/json: {schema: {$ref: '#/components/schemas/ImportCommitResult'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/ImportConflict'}
  /hash-sources:
    get:
      operationId: getHashSources
      tags: [Transfer]
      summary: List imported ROM identity data sources and their active immutable releases
      parameters: [{$ref: '#/components/parameters/Limit'}, {$ref: '#/components/parameters/Offset'}]
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Paginated hash sources, content: {application/json: {schema: {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/HashSource'}}}}}}}
  /hash-identities/{sha256}:
    get:
      operationId: getHashIdentitiesBySha256
      tags: [Transfer]
      summary: Resolve active reference metadata for an exact ROM SHA-256
      parameters:
        - {name: sha256, in: path, required: true, schema: {type: string, pattern: '^[0-9a-f]{64}$'}}
      responses:
        '401': {$ref: '#/components/responses/Error'}
        '200': {description: Zero or more source-attributed identity records, content: {application/json: {schema: {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/HashIdentity'}}}}}}}
  /hash-packs/preview:
    post:
      operationId: postHashPacksPreview
      tags: [Transfer]
      summary: Validate and classify a privacy-minimized Kino hash pack without writing
      requestBody: {$ref: '#/components/requestBodies/HashPackUpload'}
      responses:
        '200': {description: Signed hash pack preview, content: {application/json: {schema: {$ref: '#/components/schemas/HashPackPreview'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '413': {$ref: '#/components/responses/Error'}
  /hash-packs/import:
    post:
      operationId: postHashPacksImport
      tags: [Transfer]
      summary: Revalidate and atomically import the exact hash pack approved by preview
      requestBody: {$ref: '#/components/requestBodies/HashPackCommit'}
      responses:
        '200': {description: The exact immutable release was already installed; no write was needed, content: {application/json: {schema: {$ref: '#/components/schemas/HashPackImportResult'}}}}
        '201': {description: "Immutable release imported; no Game, Edition, or Artifact is created", content: {application/json: {schema: {$ref: '#/components/schemas/HashPackImportResult'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '413': {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Owner token or a server-issued device token; device scopes are enforced per route.
...
  parameters:
    ID: {name: id, in: path, required: true, schema: {type: string}, description: Opaque resource identifier; clients must not infer its shape or ordering.}
...
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 100}, description: Maximum number of records returned in this page.}
    Offset: {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}, description: Zero-based record offset in the endpoint's documented stable ordering.}
...
  responses:
...
    Error:
      description: Structured API error
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestID'}
        X-Kino-API-Version: {$ref: '#/components/headers/APIVersion'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorEnvelope'}
          example:
            error:
              code: invalid_argument
              message: request payload is invalid
              request_id: 40e39da4b75c41c28cf40b9f8de46c48
  schemas:
...
    PlatformCollection:
      allOf:
        - {$ref: '#/components/schemas/PaginatedBase'}
        - {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/Platform'}}}}
    CustomPlatformCollection:
      allOf:
        - {$ref: '#/components/schemas/PaginatedBase'}
        - {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/CustomPlatform'}}}}
...